default

+++ src/main/java/br/com/uoutec/community/ediacaran/sales/registry/ProductMetadataRegistryImp.java
@@ -2,9 +2,13 @@ package br.com.uoutec.community.ediacaran.sales.registry;
 
 import java.util.List;
 
+import javax.enterprise.context.control.ActivateRequestContext;
 import javax.inject.Inject;
 import javax.inject.Singleton;
+import javax.transaction.Transactional;
 
+import br.com.uoutec.application.security.ContextSystemSecurityCheck;
+import br.com.uoutec.community.ediacaran.sales.SalesPluginPermissions;
 import br.com.uoutec.community.ediacaran.sales.entity.ProductMetadata;
 import br.com.uoutec.community.ediacaran.sales.entity.ProductMetadataAttribute;
 import br.com.uoutec.community.ediacaran.sales.entity.ProductMetadataAttributeOption;
@@ -29,7 +33,12 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	private ProductMetadataAttributeOptionEntityAccess productMetadataAttributeOptionEntityAccess;
 	
 	@Override
+	@Transactional
+	@ActivateRequestContext
 	public void registerProductMetadata(ProductMetadata entity) throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.getRegisterPermission());
+		
 		try{
 			ProductMetadataRegistryUtil.validate(entity);
 			ProductMetadataRegistryUtil.saveOrUpdate(entity, entityAccess);
@@ -41,7 +50,12 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@Transactional
+	@ActivateRequestContext
 	public void removeProductMetadata(ProductMetadata entity) throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.getRemovePermission());
+		
 		try{
 			ProductMetadata actualEntity = ProductMetadataRegistryUtil.getActual(entity, entityAccess);
 			ProductMetadataRegistryUtil.delete(actualEntity, entityAccess);
@@ -53,7 +67,11 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@ActivateRequestContext
 	public ProductMetadata findProductMetadataById(int id) throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.getGetPermission());
+		
 		try{
 			return ProductMetadataRegistryUtil.get(id, entityAccess);
 		}
@@ -63,7 +81,11 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@ActivateRequestContext
 	public ProductMetadataSearchResult search(ProductMetadataSearch value) throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.getListPermission());
+		
 		try{
 			return ProductMetadataRegistryUtil.search(value, entityAccess);
 		}
@@ -75,7 +97,11 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	/* Attributes */
 	
 	@Override
+	@ActivateRequestContext
 	public ProductMetadataAttribute findProductMetadataAttributeById(int id) throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.getGetPermission());
+		
 		try{
 			return ProductMetadataAttributeRegistryUtil.get(id, metadataentityAccess);
 		}
@@ -85,8 +111,12 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@Transactional
+	@ActivateRequestContext
 	public void registerProductMetadataAttributes(List<ProductMetadataAttribute> attributes, ProductMetadata parent)
 			throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.getRegisterPermission());
 		
 		try {
 			for(ProductMetadataAttribute entity: attributes) {
@@ -105,8 +135,13 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@Transactional
+	@ActivateRequestContext
 	public void removeProductMetadataAttributes(List<ProductMetadataAttribute> attributes, ProductMetadata parent)
 			throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.getRemovePermission());
+		
 		try {
 			for(ProductMetadataAttribute entity: attributes) {
 				ProductMetadataAttributeRegistryUtil.delete(entity, metadataentityAccess);
@@ -120,8 +155,12 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@ActivateRequestContext
 	public List<ProductMetadataAttribute> getProductMetadataAttributes(ProductMetadata parent)
 			throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.getListPermission());
+		
 		try {
 			return ProductMetadataAttributeRegistryUtil.getByParent(parent, metadataentityAccess);
 		}
@@ -133,7 +172,11 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	/* options */
 	
 	@Override
+	@ActivateRequestContext
 	public ProductMetadataAttributeOption findProductMetadataAttributeOptionById(int id) throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.OPTIONS.getGetPermission());
+		
 		try{
 			return ProductMetadataAttributeOptionRegistryUtil.get(id, productMetadataAttributeOptionEntityAccess);
 		}
@@ -143,8 +186,12 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@Transactional
+	@ActivateRequestContext
 	public void registerProductMetadataAttributeOptions(List<ProductMetadataAttributeOption> options, ProductMetadataAttribute parent)
 			throws ProductRegistryException {
+
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.OPTIONS.getRegisterPermission());
 		
 		try {
 			for(ProductMetadataAttributeOption entity: options) {
@@ -163,8 +210,13 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@Transactional
+	@ActivateRequestContext
 	public void removeProductMetadataAttributeOptions(List<ProductMetadataAttributeOption> options, ProductMetadataAttribute parent)
 			throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.OPTIONS.getRemovePermission());
+		
 		try {
 			for(ProductMetadataAttributeOption entity: options) {
 				ProductMetadataAttributeOptionRegistryUtil.delete(entity, productMetadataAttributeOptionEntityAccess);
@@ -178,8 +230,12 @@ public class ProductMetadataRegistryImp implements ProductMetadataRegistry {
 	}
 
 	@Override
+	@ActivateRequestContext
 	public List<ProductMetadataAttributeOption> getProductMetadataAttributeOptions(ProductMetadataAttribute parent)
 			throws ProductRegistryException {
+		
+		ContextSystemSecurityCheck.checkPermission(SalesPluginPermissions.PRODUCT_METADATA.ATTRIBUTE.OPTIONS.getListPermission());
+		
 		try {
 			return ProductMetadataAttributeOptionRegistryUtil.getByParent(parent, productMetadataAttributeOptionEntityAccess);
 		}
